fix: search terms are not accepted
This project uses a set of fixed search words.  Anything different from
these will throw an error.  This is demonstrated by the 2 example
console.log statements.  One showing the result of a fixed search
word, and the other for a non-recognised search word.  Will need to
account for this, to prevent the app from failing.